Move allocation interface to Sound.SC3.Server.Monad and add MonadControl instances
Remove allocation interface from Sound.SC3.Server.Connection and move it to Sound.SC3.Server.Monad; Connection is unlikely to be used in a standalone fashion anyway, and it's clearer to only encapsulate communication in that module.

Get rid of the data-accessor dependency.

Add instances for MonadBase, MonadBaseControl and MonadTransControl, allowing to lift control operations from the base monad into ServerT. Implement fork in terms of Control.Concurrent.Lifted.fork.